+++ examples/08_reductions.cpp
@@ -5,8 +5,8 @@
 
 int main() {
   const int vector_length = 2000;
-  std::vector<double> x_host(vector_length, 1.0);
-  std::vector<double> y_host(vector_length, 1.0);
+  std::vector<float> x_host(vector_length, 1.0);
+  std::vector<float> y_host(vector_length, 1.0);
 
   // Create a sycl::queue using the default device selector
   sycl::device sycl_device{sycl::default_selector()};
@@ -14,11 +14,11 @@ int main() {
   sycl::queue sycl_queue{sycl_context, sycl_device};
 
   // Allocate vectors on the device
-  double* x =
-      sycl::malloc_device<double>(vector_length, sycl_device, sycl_context);
-  double* y =
-      sycl::malloc_device<double>(vector_length, sycl_device, sycl_context);
-  double* xdoty = sycl::malloc_device<double>(1, sycl_device, sycl_context);
+  float* x =
+      sycl::malloc_device<float>(vector_length, sycl_device, sycl_context);
+  float* y =
+      sycl::malloc_device<float>(vector_length, sycl_device, sycl_context);
+  float* xdoty = sycl::malloc_device<float>(1, sycl_device, sycl_context);
 
   // Copy from the host to the device.
   sycl::event copy_x = sycl_queue.copy(x_host.data(), x, x_host.size());
@@ -48,11 +48,11 @@ int main() {
   });
 
   // Copy result from device to host; synchronize;
-  double xdoty_host;
+  float xdoty_host;
   sycl_queue.copy(xdoty, &xdoty_host, 1, {dotp_kernel}).wait();
 
   // Verify the results.
-  if (static_cast<double>(vector_length) != xdoty_host) {
+  if (static_cast<float>(vector_length) != xdoty_host) {
     std::cout << "Verification failed!\n";
     return EXIT_FAILURE;
   }
